+++ packages/preact/vitest.config.js
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vitest/config';
 import { createRequire } from 'node:module';
+import preact from '@preact/preset-vite';
 
 const require = createRequire(import.meta.url);
 
@@ -16,7 +17,11 @@ export default defineConfig({
       },
     ],
   },
+  plugins: [preact()],
   test: {
+    transformMode: {
+      web: [/\.[jt]sx$/],
+    },
     environment: 'jsdom',
   },
 });