+++ packages/reporter-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @felte/reporter-solid
 
+## 1.2.9
+
+### Patch Changes
+
+- Updated dependencies [0a99410]
+  - @felte/common@1.1.8
+
 ## 1.2.8
 
 ### Patch Changes